+++ plugins/content-manager/src/main/java/com/wazuh/contentmanager/engine/service/EngineService.java
@@ -46,6 +46,4 @@ public interface EngineService {
      *     {"hash":"...","updating":false}}).
      */
     RestResponse getIocState();
-
-    RestResponse deleteLogtest();
 }