+++ plugins/content-manager/src/test/java/com/wazuh/contentmanager/engine/service/EngineServiceImplTests.java
@@ -170,26 +170,6 @@ public class EngineServiceImplTests extends OpenSearchTestCase {
                                                         .equals("Test Rule")));
     }
 
-    /** Tests the deleteLogtest operation for a successful (200) response. */
-    public void testDeleteLogtest200() {
-        RestResponse expected = new RestResponse("OK", 200);
-        when(this.socket.sendRequest(eq(EngineServiceImpl.LOGTEST), eq("DELETE"), any(JsonNode.class)))
-                .thenReturn(expected);
-
-        RestResponse actual = this.engine.deleteLogtest();
-        assertEquals(expected, actual);
-    }
-
-    /** Tests the deleteLogtest operation for an internal server error (500) response. */
-    public void testDeleteLogtest500() {
-        RestResponse expected = new RestResponse("Engine Socket Error", 500);
-        when(this.socket.sendRequest(eq(EngineServiceImpl.LOGTEST), eq("DELETE"), any(JsonNode.class)))
-                .thenReturn(expected);
-
-        RestResponse actual = this.engine.deleteLogtest();
-        assertEquals(expected, actual);
-    }
-
     /** Tests that getIocState sends a GET request to the correct endpoint. */
     public void testGetIocState200() {
         RestResponse expected = new RestResponse("{\"hash\":\"abc\",\"updating\":false}", 200);